+++ OpenGL/src/maths/Maths.h
@@ -2,7 +2,7 @@
 
 namespace Maths {
 
-	constexpr float PI = 3.14159265358979323846;
+	constexpr double PI = 3.14159265358979323846;
 	constexpr float PI_OVER_ONE_EIGHTY = PI / 180;
 	constexpr float ONE_EIGHTY_OVER_PI = 180 / PI;
